Motiondetection fixed so it doesn't give motion on first frame, tests rewritten (#52)

+++ server/tests/tests.cpp
@@ -1,10 +1,11 @@
 
 // Testfil där alla enhetstester för backend kommer skrivas mha catch2 framework.
 // Här kan man skriva test cases på sättet du ser nedan.
-// Två executables skapas när vera körs, en som kör hemsidan och en som
-// Kör tester. För att det ska kompilera korrekt behöver du lägga till alla .cpp filer som
+// För att det ska kompilera korrekt behöver du lägga till alla .cpp filer som
 // ditt test använder i server/CMakeLists.txt add_executables(tests...) (rad 96) Observera ej
 // server/tests/Cmakelists
+
+// Testfilen körs genom python3 run_tests.py från terminalen
 
 #include "http_server.h"
 #define CATCH_CONFIG_MAIN  // This creates a main() function automatically
@@ -16,6 +17,8 @@
 
 #include <filesystem>
 #include <fstream>
+
+/* Ljudanalys testning */
 
 TEST_CASE("extract creates a .wav file from video input", "[extract]") {
   std::string video_file = "sample_video.mp4";
@@ -29,8 +32,11 @@ TEST_CASE("extract creates a .wav file from video input", "[extract]") {
 
   user_path.append("sample_video.mp4");
 
-  std::filesystem::copy_file(path, user_path);
+  if(!std::filesystem::exists(user_path)){
+    std::filesystem::copy_file(path, user_path);
+  }
 
+  // Verifiera att filen skapats korrekt i user_pathen
   REQUIRE(std::filesystem::exists(user_path));
 
   extract(video_file, test_session);
@@ -45,16 +51,6 @@ TEST_CASE("extract creates a .wav file from video input", "[extract]") {
   std::filesystem::remove(user_path);
 }
 
-TEST_CASE("Kastar undantag om ROI är utanför bilden", "[error]") {
-  std::vector<int> invalid_coords = {500, 2000, 500, 2000};  // Medvetet för stort ROI
-  REQUIRE_THROWS_AS(analyse_video("", invalid_coords), std::invalid_argument);
-}
-
-TEST_CASE("Kastar undantag om ROI börjar utanför bilden", "[error]") {
-  std::vector<int> invalid_coords = {-500, 500, -500,
-                                     500};  // Medvetet ROI som börjar utanför bilden
-  REQUIRE_THROWS_AS(analyse_video("", invalid_coords), std::invalid_argument);
-}
 TEST_CASE("db_analysis upptäcker högljudda segment från dB-textfil", "[audio-analysis]") {
   // Skapa testsession
   UserSession test_session{"test"};
@@ -109,4 +105,45 @@ TEST_CASE("db_analysis upptäcker högljudda segment från dB-textfil", "[audio-
   // Rensa upp efter testet
   std::filesystem::remove(full_input_path);
   std::filesystem::remove(output_path);
+}
+
+
+/* Videoanalys testning */
+
+TEST_CASE("Testar att analyse_video() ger en rimlig output, rätt mängd rörelsetillfällen"){
+  std::string video_file = "sample_video.mp4";
+
+  std::filesystem::path path = getTestingDir();
+  path.append(video_file);
+
+  UserSession test_session = UserSession{"test"};
+  std::string user_path = getUserMediaDir(test_session);
+
+  user_path.append("sample_video.mp4");
+
+ if(!std::filesystem::exists(user_path)){
+    std::filesystem::copy_file(path, user_path);
+  }
+
+  std::vector<int> coords = {1, 200, 1, 200};
+  std::vector<int> result = analyse_video(user_path, coords);
+
+  REQUIRE(result.size() == 4);
+
+}
+
+TEST_CASE("Kastar undantag om filväg ej hittas", "[error]"){
+  std::vector<int> valid_coords = {1, 200, 1, 200};
+  REQUIRE_THROWS_AS(analyse_video("", valid_coords), std::invalid_argument);
+}
+
+TEST_CASE("Kastar undantag om ROI är utanför bilden", "[error]") {
+  std::vector<int> invalid_coords = {500, 2000, 500, 2000};  // Medvetet för stort ROI
+  REQUIRE_THROWS_AS(analyse_video("sample_video.mp4", invalid_coords), std::invalid_argument);
+}
+
+TEST_CASE("Kastar undantag om ROI börjar utanför bilden", "[error]") {
+  std::vector<int> invalid_coords = {-500, 500, -500,
+                                     500};  // Medvetet ROI som börjar utanför bilden
+  REQUIRE_THROWS_AS(analyse_video("sample_video.mp4", invalid_coords), std::invalid_argument);
 }
